Add files via upload
Created this project in Java and all these classes are used to make the entire images work and for audio files. Mainly the Bitmap is for images as .bmp and audio files are in .wav files for the program to actually work as intended and changes the way the user chooses the audio or images to look like.